LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/interfaces-sw.https-expected.txt


Harness Error (TIMEOUT), message = null

PASS Interfaces and attributes in ServiceWorkerGlobalScope 
TIMEOUT test setup (cache creation) Test timed out
FAIL Event constructors assert_equals: Default FetchEvent.clientId should be null expected (object) null but got (string) ""
PASS xhr is not exposed 
PASS ServiceWorkerGlobalScope interface: existence and properties of interface object 
PASS ServiceWorkerGlobalScope interface object length 
PASS ServiceWorkerGlobalScope interface object name 
PASS ServiceWorkerGlobalScope interface: existence and properties of interface prototype object 
TIMEOUT ServiceWorkerGlobalScope interface: internal [[SetPrototypeOf]] method of interface prototype object - setting to a new value via Object.setPrototypeOf should throw a TypeError Test timed out
